+++ canvas-dao/canvas-dao-api/src/main/java/org/renci/hearsay/canvas/clinbin/dao/model/BinResultsFinalDiagnostic.java
@@ -108,7 +108,7 @@ public class BinResultsFinalDiagnostic implements Persistable {
     private String accNum;
 
     @Column(name = "max_allele_freq")
-    private Float maxAlleleFreq;
+    private Double maxAlleleFreq;
 
     @Lob
     @Column(name = "refallele")
@@ -133,19 +133,19 @@ public class BinResultsFinalDiagnostic implements Persistable {
     private Integer depth;
 
     @Column(name = "qd")
-    private Float qd;
+    private Double qd;
 
     @Column(name = "read_pos_rank_sum")
-    private Float readPosRankSum;
+    private Double readPosRankSum;
 
     @Column(name = "frac_reads_with_dels")
-    private Float fracReadsWithDels;
+    private Double fracReadsWithDels;
 
     @Column(name = "hrun")
     private Integer hrun;
 
     @Column(name = "strand_score")
-    private Float strandScore;
+    private Double strandScore;
 
     @Column(name = "ref_depth")
     private Integer refDepth;
@@ -157,7 +157,7 @@ public class BinResultsFinalDiagnostic implements Persistable {
     private Boolean homozygous;
 
     @Column(name = "genotype_qual")
-    private Float genotypeQual;
+    private Double genotypeQual;
 
     @Lob
     @Column(name = "tier")
@@ -370,11 +370,11 @@ public class BinResultsFinalDiagnostic implements Persistable {
         this.accNum = accNum;
     }
 
-    public Float getMaxAlleleFreq() {
+    public Double getMaxAlleleFreq() {
         return maxAlleleFreq;
     }
 
-    public void setMaxAlleleFreq(Float maxAlleleFreq) {
+    public void setMaxAlleleFreq(Double maxAlleleFreq) {
         this.maxAlleleFreq = maxAlleleFreq;
     }
 
@@ -442,27 +442,27 @@ public class BinResultsFinalDiagnostic implements Persistable {
         this.depth = depth;
     }
 
-    public Float getQd() {
+    public Double getQd() {
         return qd;
     }
 
-    public void setQd(Float qd) {
+    public void setQd(Double qd) {
         this.qd = qd;
     }
 
-    public Float getReadPosRankSum() {
+    public Double getReadPosRankSum() {
         return readPosRankSum;
     }
 
-    public void setReadPosRankSum(Float readPosRankSum) {
+    public void setReadPosRankSum(Double readPosRankSum) {
         this.readPosRankSum = readPosRankSum;
     }
 
-    public Float getFracReadsWithDels() {
+    public Double getFracReadsWithDels() {
         return fracReadsWithDels;
     }
 
-    public void setFracReadsWithDels(Float fracReadsWithDels) {
+    public void setFracReadsWithDels(Double fracReadsWithDels) {
         this.fracReadsWithDels = fracReadsWithDels;
     }
 
@@ -474,11 +474,11 @@ public class BinResultsFinalDiagnostic implements Persistable {
         this.hrun = hrun;
     }
 
-    public Float getStrandScore() {
+    public Double getStrandScore() {
         return strandScore;
     }
 
-    public void setStrandScore(Float strandScore) {
+    public void setStrandScore(Double strandScore) {
         this.strandScore = strandScore;
     }
 
@@ -506,11 +506,11 @@ public class BinResultsFinalDiagnostic implements Persistable {
         this.homozygous = homozygous;
     }
 
-    public Float getGenotypeQual() {
+    public Double getGenotypeQual() {
         return genotypeQual;
     }
 
-    public void setGenotypeQual(Float genotypeQual) {
+    public void setGenotypeQual(Double genotypeQual) {
         this.genotypeQual = genotypeQual;
     }
